fix timestamp's type

+++ ziyan/ziyan_main.py
@@ -133,7 +133,7 @@ class Handler(object):
                 if self.unit == 's':
                     timestamp = processed_dict.get('timestamp', pendulum.now().int_timestamp)
                 else:
-                    timestamp = processed_dict.get('timestamp', pendulum.now().float_timestamp * 1000000)
+                    timestamp = processed_dict.get('timestamp', int(pendulum.now().float_timestamp * 1000000))
 
                 update_dict = {'fields': fields, 'timestamp': timestamp}
 
@@ -166,7 +166,7 @@ class Handler(object):
             if self.unit == 's':
                 timestamp = processed_dicts.get('timestamp', pendulum.now().int_timestamp)
             else:
-                timestamp = processed_dicts.get('timestamp', pendulum.now().float_timestamp * 1000000)
+                timestamp = processed_dicts.get('timestamp', int(pendulum.now().float_timestamp * 1000000))
 
             update_dict = {'fields': fields, 'timestamp': timestamp}
